Automatically populate date/idmap.json with UIDs
Rather than having to manually copy UIDs into the id map, this will load
the JSON (defaulting to an empty list) and append the current UID to a
new entry in the list when it doesn't exist, then rewrites the file.

Also removes the ability to use UIDs directly in admin/user lists
because it makes no sense when usernames will be auto-mapped.
The `show_uids` setting is however retained as `detailed_denials`.