Cleanup select for mobile, fix IPv6 addresses
On (some?) mobile browsers the select's `size` option has no effect so
this cleans it up a bit, moving the title to a fake option so it doesn't
make the desktop version look stupid with two headings.

If you get an IPv6 address, it could also be too long and make the whole
page need scrolling, so that's also fixed with a few extra styles.