Trust local proxies to get reliable log IPs
This doesn't affect the code side of things (X_FORWARDED_FOR seems to be
passed through from Traefik regardless), but it does get Caddy to report
the correct IP (rather than that of the Traefik proxy) in access logs.

Despite Cloudflare also being in the loop, without having Caddy's strict
option set it will read left-to-right, so as long as there's a local
proxy as well, there's no need to add Cloudflare's ranges here.