Get routes working properly
Although Caddy directive order matters to an extent, it doesn't when
multiple lines/blocks match the same file. When that happens, the result
is essentially a lottery. Usually, PHP catches it all so we get no CSS.
Sometimes. Sometimes when you start the container it would get lucky!

This commit fixes it once and for all, scoping both PHP *and* the file
server because otherwise that would display all PHP files as plain text
for the same reasons as before. To solve the .php redirect issue there's
just a bunch of `redir` directives. Regex seems to just result in NOP.