Automatically trust local Caddy proxy when found
If we detect a Caddy server, and the client IP is in one of the default
Docker networks, and we can also resolve Caddy by the default container
name (i.e. "frontend"), then we set its IP in the proxies list.

Should (somewhat safely, assuming some sort of network isolation) deal
with the issue of whitelisting the stock caddy proxy without having to
hardcode a specific static IP for each of the compose services.